Rewrite number.....10? Added Countries to the tanks. Merged listeners and handlers into battlefield class.

+++ common/src/main/java/musician101/minetanks/common/util/AbstractScoreboard.java
@@ -32,5 +32,5 @@ public abstract class AbstractScoreboard<Scoreboard>
 
     public abstract void playerDeath(UUID playerId);
 
-    public abstract void setPlayerHealth(UUID playerId, int health);
+    public abstract void setPlayerHealth(UUID playerId, double health);
 }
